Add a .gitattributes file
The reason for adding this file initially is that I was using vscode
and it kept kicking my cursor to the top of the file.

I initially gave up and used Vim, but the hinting in vscode is too
good to pass up, so I did some research.

I found found somehwere (probably stack overflow) that if I change the
line endings of a file from CRLF to LF it fixes the problem. I did
this in vscode and it worked! Yay!

But then whenever I checked out a different branch, git kicked all
the files back to CRLF. Boo.

I remembered that I could add this .gitattributes file to tell git:
"Hey files that match this pattern are text files and should have
LF line endings, OK?"

Git graceously agreed to keep my line endings so now we have this
file.