Update TESTING doc to reflect new response func
The `mockServer()` method now returns a body _and_ a status code.

The reason for this change was that with the addition of
`authWrappedRequest()` we need to be able to test that the API Client
acts as expected when the status code is a 401 or 403.

We previously assumed all requests were "good" and returned a 200,
but this allows us to perform much more expressive tests and
especially empowers us to create more negative tests.

This commit isn't that change. I forgot to update the TESTING doc,
so this is the docs change instead.