Fixing auth wrapped request.
the authWrappedRequest method had a problem where POST requests
would have an empty body during the retry.

This fixes the issue by changing the interface to makeRequest
and authWrappedRequest so the former gets all the inputs to
http.NewRequest and the latter gets a request and a "backup request"
to use as the fallback.

This isn't an _ideal_ solution, but it fixes the initial bug and
while creating two request objects isn't great, I'm going to deem
the sub-optimal solution OK for now. The complexity of creating
two request objects for every request is basically nothing
(in my opionion) next to the time burned making two API calls
(c.Authenticate(), request.Do()).

I think keeping the simple interface (never needing to explicitly
call Authenticate or check for unauthorized status codes) is worth
more than what is lost creating duplicate reuqest objects.

If this becomes a problem later on, we can discuss longer-term
solutions.